standardized enroll error messages, 2- showing error message when email already exists, email is invalid or textfields has less length than minLength
