+++ src/pages/Enroll/index.js
@@ -1,114 +1,7 @@
-import { useContext, useState } from "react";
-import { toast } from "react-toastify";
-import { useHistory } from "react-router-dom";
-
-import AuthLayout from "../../layouts/Auth";
-
-import Input from "../../components/Form/Input";
-import Button from "../../components/Form/Button";
-import { Row, Title, Label } from "../../components/Auth";
-import Link from "../../components/Link";
-
-import EventInfoContext from "../../contexts/EventInfoContext";
-import UserContext from "../../contexts/UserContext";
-
-import useApi from "../../hooks/useApi";
+import EnrollForm from "../../components/EnrollForm/EnrollForm";
 
 export default function Enroll() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [loadingEnroll, setLoadingEnroll] = useState(false);
-
-  const { setUserData } = useContext(UserContext);
-  const history = useHistory();
-
-  const api = useApi();
-
-  const { eventInfo } = useContext(EventInfoContext);
-
-  const signIn = (email, password) => {
-    api.auth
-      .signIn(email, password)
-      .then((response) => {
-        setUserData(response.data);
-      })
-      .catch(() => {
-        history.push("sign-in");
-      });
-  };
-
-  function submit(event) {
-    event.preventDefault();
-
-    if (password !== confirmPassword) {
-      toast("As senhas devem ser iguais!");
-    } else {
-      setLoadingEnroll(true);
-      api.user
-        .signUp(email, password)
-        .then((response) => {
-          toast("Inscrito com sucesso!");
-          signIn(email, password);
-        })
-        .catch((error) => {
-          if (error.response) {
-            for (const detail of error.response.data.details) {
-              toast(detail);
-            }
-          } else {
-            toast("Não foi possível conectar ao servidor!");
-          }
-        })
-        .then(() => {
-          setLoadingEnroll(false);
-        });
-    }
-  }
-
   return (
-    <AuthLayout background={eventInfo.backgroundImage}>
-      <Row>
-        <img src={eventInfo.logoImage} alt="Event Logo" />
-        <Title>{eventInfo.eventTitle}</Title>
-      </Row>
-      <Row>
-        <Label>Inscrição</Label>
-        <form onSubmit={submit}>
-          <Input
-            label="E-mail"
-            type="text"
-            fullWidth
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-          <Input
-            label="Senha"
-            type="password"
-            fullWidth
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-          />
-          <Input
-            label="Repita sua senha"
-            type="password"
-            fullWidth
-            value={confirmPassword}
-            onChange={(e) => setConfirmPassword(e.target.value)}
-          />
-          <Button
-            type="submit"
-            color="primary"
-            fullWidth
-            disabled={loadingEnroll}
-          >
-            {loadingEnroll ? "Carregando..." : "Inscrever"}
-          </Button>
-        </form>
-      </Row>
-      <Row>
-        <Link to="/sign-in">Já está inscrito? Faça login</Link>
-      </Row>
-    </AuthLayout>
+    <EnrollForm />
   );
 }
